refactor: ui 패키지 websokcet 내부 패키지로 변경
