+++ backend/src/main/java/coffeeshout/global/exception/WebSocketExceptionHandler.java
@@ -1,6 +1,6 @@
 package coffeeshout.global.exception;
 
-import coffeeshout.global.ui.WebSocketResponse;
+import coffeeshout.global.websocket.ui.WebSocketResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.messaging.handler.annotation.Header;
 import org.springframework.messaging.handler.annotation.MessageExceptionHandler;
