+++ backend/src/main/java/coffeeshout/global/websocket/ui/WebSocketResponse.java
@@ -1,4 +1,4 @@
-package coffeeshout.global.ui;
+package coffeeshout.global.websocket.ui;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
